Change pointers on smart pointers

+++ Coordinates.cpp
@@ -2,22 +2,22 @@
 
 void Coordinates::SetCoordinates(double _x, double _y, double _z)
 {
-	x.push_back(static_cast<float>( _x));
-	y.push_back(static_cast<float>( _y));
-	z.push_back(static_cast<float>( _z));
+	x.push_back(static_cast<GLdouble>(_x));
+	y.push_back(static_cast<GLdouble>(_y));
+	z.push_back(static_cast<GLdouble>(_z));
 }
 
 void Coordinates::SetCoordinates(double _x, double _y)
 {
-	x.push_back(static_cast<float>( _x));
-	y.push_back(static_cast<float>( _y));
+	x.push_back(static_cast<GLdouble>(_x));
+	y.push_back(static_cast<GLdouble>(_y));
 }
 
 void Coordinates::SetCoordinates(double _x)
 {
-	x.push_back(static_cast<float>( _x));
-	y.push_back(static_cast<float>( _x));
-	z.push_back(static_cast<float>( _x));
+	x.push_back(static_cast<GLdouble>(_x));
+	y.push_back(static_cast<GLdouble>(_x));
+	z.push_back(static_cast<GLdouble>(_x));
 }
 
 void Coordinates::Clear()
